+++ net.sourceforge.pmd.eclipse.plugin.test/src/main/java/net/sourceforge/pmd/eclipse/internal/ResourceUtil.java
@@ -7,8 +7,14 @@ package net.sourceforge.pmd.eclipse.internal;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.io.OutputStream;
+import java.io.Reader;
 import java.nio.file.Files;
+
+import org.eclipse.core.resources.IFile;
+import org.eclipse.core.resources.IProject;
+import org.eclipse.core.runtime.CoreException;
 
 public final class ResourceUtil {
     private ResourceUtil() {
@@ -30,5 +36,23 @@ public final class ResourceUtil {
                 count = in.read(buffer);
             }
         }
+    }
+
+    public static String getResourceAsString(IProject project, String resourceName) throws IOException, CoreException {
+        IFile file = project.getFile(resourceName);
+        String charset = file.getCharset();
+        char[] buffer = new char[1024];
+        StringBuilder result = new StringBuilder();
+        try (Reader in = new InputStreamReader(file.getContents(), charset)) {
+            int count = in.read(buffer);
+            while (count > -1) {
+                result.append(buffer, 0, count);
+                count = in.read(buffer);
+            }
+            if (count > -1) {
+                result.append(buffer, 0, count);
+            }
+        }
+        return result.toString();
     }
 }